add winner method and bug fixes
Created a method called winner. It distributes the daily prize to the winner/s. Moreover, bug fixes was performed: Ensured the differentiation of the total number of machines (stored in the total_machines attribute) and the total number of active machines (stored in the active_machines attribute). Lastly, adjustments were made to the mining_player attribute, ensuring that only active users are taken into account.