Create bankrupt_status attribute and bankrupt_check method
bankrupt_status is used to check the bankruptcy status of the user. bankrupt_check method is used to check the user’s capital, ensuring that if the balance is negative, SDPA coin will automatically be sold from their SDPA balance to cover the negative balance. If the negative balance cannot be covered, the user will be declared bankrupt.